+++ secret-hitler/program/programs/secret-hitler/src/state/nomination.rs
@@ -2,8 +2,8 @@ use anchor_lang::prelude::*;
 
 #[account]
 pub struct Nomination {
-    pub voters_index: Vec<usize>,
-    pub nominee_index: usize,
+    pub voters_index: Vec<u64>,
+    pub nominee_index: u64,
     pub nein: u8,
     pub ja: u8,
     pub bump: u8,
@@ -14,7 +14,7 @@ impl Space for Nomination {
 }
 
 impl Nomination {
-    pub fn init(&mut self, index: usize, president_index: usize, bump: u8) {
+    pub fn init(&mut self, index: u64, president_index: u64, bump: u8) {
         self.voters_index = vec![president_index];
         self.nominee_index = index;
         self.nein = 0;
